ALL TOML VALUES ARE ORDERED NOW, FYI.
We needed to preserve order when passing one map from place to place in
the build system, and to do that, we had to turn on ordering for all
toml maps everyhwere in our entire dependency graph, due to an
incredibly misguided design decision in the toml crate.

If you've bisected an issue to this commit please talk to me.